Help message now in InfoRequestFactory
Set introduced to avoid duplicating
